refactor, trainer is able to build OC-SVM
